# docs/swagger/collections.yaml
openapi: 3.0.0
servers:
  - description: ListPocket Development Server
    url: http://localhost:9000/api
info:
  version: "1.0.0"
  title: ListPocket
  description: The API collection for listpocket
  license:
    name: AGPL-3.0 license
    url: https://github.com/knadh/listmonk/blob/master/LICENSE
...
paths:
...
  /import/subscribers:
    get:
      description: returns import status.
      operationId: getImportSubscribers
      tags:
        - Import
      responses:
        "200":
          description: import status
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ImportStatus"

    post:
      description: handles the uploading and bulk importing of a ZIP file of one or more CSV files.
      operationId: importSubscribers
      tags:
        - Import
      requestBody:
        description: uploads and bulk imports of compressed CSV files
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                params:
                  type: string
                  description: JSON string containing import parameters for more detail https://listmonk.app/docs/apis/import/#params-json-string
                  example: '{"mode":"subscribe", "subscription_status":"confirmed", "delim":",", "lists":[1, 2], "overwrite": true}'
                file:
                  type: string
                  format: binary
                  description: File for upload.
      responses:
        "200":
          description: updated import status
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ImportStatus"

    delete:
      description: sends a stop signal to the importer.
      operationId: stopImportSubscribers
      tags:
        - Import
      responses:
        "200":
          description: response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ImportStatus"
...
  /templates:
    get:
      description: handles retrieval of templates
      operationId: getTemplates
      tags:
        - Templates
      parameters:
        - in: query
          name: no_body
          description: boolean flag for response with/without body
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Template"
    post:
      description: handles template creation
      operationId: createTemplate
      tags:
        - Templates
      requestBody:
        description: new template info
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTemplate"
      responses:
        "200":
          description: response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Template"
...
  /templates/preview:
    post:
      description: get the HTML preview of a template.
      operationId: previewTemplate
      tags:
        - Templates
      requestBody:
        required: true
        description: template parameters
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                template_type:
                  description: type of template
                  type: string
                body:
                  description: template body
                  type: string
      responses:
        "200":
          description: response
          content:
            text/html:
              schema:
                type: string
                example: <p>Hi there</p>
...
components:
  schemas:
...
    ImportStatus:
      type: object
      properties:
        data:
          type: object
          properties:
            name:
              type: string
            total:
              type: integer
            imported:
              type: integer
            status:
              type: string
...
    Template:
      type: object
      properties:
        id:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
        name:
          type: string
        body:
          type: string
        body_source:
          type: string
        subject:
          type: string
        type:
          type: string
        is_default:
          type: boolean
    NewTemplate:
      type: object
      required:
        - name
        - type
        - body
      properties:
        name:
          type: string
          description: Name of the template
        type:
          type: string
          enum: [campaign, campaign_visual, tx]
          description: Type of the template
        subject:
          type: string
          description: Subject line for the template (only for tx)
        body_source:
          type: string
          description: JSON source for the email-builder template (only for campaign_visual)
        body:
          type: string
          description: HTML body of the template
...